wificond: get wiphy index using base interface name
When there are multiple wiphy objects registered (for example,
with both wifi and wigig drivers loaded), the current code
returns the wiphy index of the last registered driver. This
means it can return the wrong wiphy index if the wlan driver
registered first. Also if the drivers are loaded/unloaded on-demand
the registration order can change at runtime, the code can
return a different wiphy index every time.
Fix this by using a base interface name (such as wlan0) and
only returning the wiphy index of the wiphy object containing
this interface.

Change-Id: I20c4fc7465a37ab21cba76339d78da3a92112d35