wifi: Check for sched scan's relative_rssi support to use it
Current behavior of the framework is to set the attributes of
relative_rssi by default in the sched scan request.
Kernel returns a failure if the host driver does not support
this capability , but the framework sends the respective
attributes.
Hence,check the driver support of this feature before accessing
its functionality.

Change-Id: I51001ac1c167e2fd7ad0508b0e1e02fb11ad6677
CRs-fixed: 2308247