Add a Wi-Fi recovery (assert) when 10 successive scans fail with ENETDOWN.
This commit adds a recovery mechanism when the successive 10 scans
fail with the reason ENETDOWN.
This situation can be reproduced when an attempt to set the mac
randomization fails while the Wi-Fi driver is undergoing a sub system
restart. ( user selects a different profile when in connected state ,
while triggering the SSR explicitly, during this new profile connection,
a set mac randomization fails and thus the interface is never brought up).

Once this situation occurs , the interface is down and all the scan attempts
fail with the reason ENETDOWN. This only recovery for this is a Wi-Fi restart.

Address this situation by triggering an ASSERT if all the successive scans / PNO
fail with the reason ENETDOWN for 10 iterations.

Change-Id: I9c0a2d0ec6129f6bde6b5944b918c58de749cad7
CRs-Fixed: 2530049