Remove libutils.a whole_static_lib from libwificond_event_loop.a
Bug: http://b/197965342

All users of this library depend on libutils.so.  Having this dependency
breaks build of wificond binary with upstream LLD.  This is because
wificond depends on libbinder which has undefined symbols provided by
libutils.so.  If libutils.a is included in the link command, the linker
tries to redefine these undefined symbols by fetching symbols from
libutils.a.

Test: build with aosp/1809741 which has lld with the above behavior.
Test: mmma system/connectivity/wificond
Change-Id: I0b40d3ae8864ae22f088935a3d7168486d833aa9
Signed-off-by: Jason Edson <jaysonedson@gmail.com>