Update the band to wiphy index map on country code change
On certain driver/firmware implementations, in world mode,
5GHz BAND information is not advertised to kernel during
the wiphy registration. Only when the country code is updated,
5GHz BAND support is updated in the driver/kernel. So
call UpdateBandWiphyIndexMap() on receiving country code
change events from kernel/driver. This is needed for wificond
to return the right channel list corresponding to the BAND.

Bug: 206337065
Test: Manual - Set country code through shell command and confirmed
from logs that the channel list is updated correctly according to the country code.

Change-Id: I528c4d6a20753b7295173a4fdb4a17285b0cd682