wifi: extend netlink command timeout and add retry mechanism
Netlink is not a reliable protocol. It might face command drop or
command timeout (base on the timer) when executing netlink command.
There is a corner case which network command timeout when turning on
Soft Ap. It is caused by kernel busy. When timeout happened, it is
easy to face "next" request command drop because buffer is full
since there was a old timeout packet in the buffer.

So extend timeout from 1 second to 4 second base on local test result.
Add retry mechanism for command drop case to recover the timeout case.

Bug: 236662100
Test: Manual test(Stress off/on softap quickly)
Change-Id: I43c8240da6e25c3e9498a4a9f0f3ecefe86da44f